fix: agent registration logic

+++ src/auth/auth.service.ts
@@ -48,9 +48,7 @@ export class AuthService {
       throw new ConflictException('User already exists')
     }
 
-    const agentRegistration = await this.repository.pendaftaranAgen.findByEmail(
-      email
-    )
+    const agentRegistration = await this.repository.pendaftaranAgen.findByEmail(email)
     if (agentRegistration) {
       if (agentRegistration.statusPendaftaran === 'DIAJUKAN') {
         throw new ConflictException('Agent registration is being processed')
@@ -66,7 +64,7 @@ export class AuthService {
       userRole = 'TRAVELER'
     }
 
-    if (userRole === 'AGEN') {
+    if (userRole === 'AGEN' && agentRegistration) {
       await this.repository.pendaftaranAgen.deleteByEmail(email)
     }
